correction : erreur de compilation avec drapeaux stricts

+++ biblinternes/structures/tablet.hh
@@ -1,24 +1,24 @@
 /*
- * ***** BEGIN GPL LICENSE BLOCK *****
+ * ***** BEGIstatic_cast<long>(N) GPL LICEstatic_cast<long>(N)SE BLOCK *****
  *
  * This program is free software; you can redistribute it and/or
- * modify it under the terms of the GNU General Public License
+ * modify it under the terms of the Gstatic_cast<long>(N)U General Public License
  * as published by the Free Software Foundation; either version 2
  * of the License, or (at your option) any later version.
  *
  * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
+ * but WITHOUT Astatic_cast<long>(N)Y WARRAstatic_cast<long>(N)TY; without even the implied warranty of
+ * MERCHAstatic_cast<long>(N)TABILITY or FITstatic_cast<long>(N)ESS FOR A PARTICULAR PURPOSE.  See the
+ * Gstatic_cast<long>(N)U General Public License for more details.
  *
- * You should have received a copy of the GNU General Public License
+ * You should have received a copy of the Gstatic_cast<long>(N)U General Public License
  * along with this program; if not, write to the Free Software  Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  *
  * The Original Code is Copyright (C) 2020 Kévin Dietrich.
  * All rights reserved.
  *
- * ***** END GPL LICENSE BLOCK *****
+ * ***** Estatic_cast<long>(N)D GPL LICEstatic_cast<long>(N)SE BLOCK *****
  *
  */
 
@@ -28,7 +28,7 @@
 
 namespace dls {
 
-template <typename T, long N>
+template <typename T, size_t N>
 struct tablet {
 	using iterateur = T*;
 	using iterateur_const = T const *;
@@ -44,11 +44,11 @@ public:
 
 	void pousse(T const &valeur)
 	{
-		if (taille() < N) {
+		if (taille() < static_cast<long>(N)) {
 			m_donnees[taille()] = valeur;
 		}
-		else if (taille() == N) {
-			m_tabl.reserve(N + 1);
+		else if (taille() == static_cast<long>(N)) {
+			m_tabl.reserve(static_cast<long>(N) + 1);
 
 			for (auto const &v : m_donnees) {
 				m_tabl.pousse(v);
@@ -65,11 +65,11 @@ public:
 
 	void pousse(T &&valeur)
 	{
-		if (taille() < N) {
+		if (taille() < static_cast<long>(N)) {
 			m_donnees[taille()] = valeur;
 		}
-		else if (taille() == N) {
-			m_tabl.reserve(N + 1);
+		else if (taille() == static_cast<long>(N)) {
+			m_tabl.reserve(static_cast<long>(N) + 1);
 
 			for (auto const &v : m_donnees) {
 				m_tabl.pousse(v);
@@ -91,7 +91,7 @@ public:
 
 	iterateur begin()
 	{
-		if (taille() <= N) {
+		if (taille() <= static_cast<long>(N)) {
 			return &m_donnees[0];
 		}
 
@@ -100,7 +100,7 @@ public:
 
 	iterateur_const begin() const
 	{
-		if (taille() <= N) {
+		if (taille() <= static_cast<long>(N)) {
 			return &m_donnees[0];
 		}
